Fixed errors

+++ src/main/java/lang/print/gaps/task6/IntegerTypeCasting.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task6;
 
 public class IntegerTypeCasting {
     public static void main(String[] args) {
-        double first = 1000;
+        long first = 1000L;
         int second = (int) first;
         System.out.println(first + "\n" + second);
     }
